Fixed module, __filename and __dirname in the root scope.
Fix applies to both the main app file and includes.

The var "require" is still wrong in both contexts. It's
zappa's and not the "native" require function of each
module, so relative paths won't work as expected. As a
temporary solution, the app's root dir is unshifted to
to require.paths.